mcp/server/requirements.txt
fastmcp==3.2.4
pyyaml>=6.0
